Break out stats helpers into their own method
